This patch addresses several issues:
* NUTCH-415 - Generator should mark selected records in CrawlDb.
  Due to increased resource consumption this step is optional.
  Application-level locking has been added to prevent concurrent
  modification of databases.

* NUTCH-416 - CrawlDatum status and CrawlDbReducer refactoring. It is
  now possible to correctly update CrawlDb from multiple segments.
  Introduce new status codes for temporary and permanent
  redirection.

* NUTCH-322 - Fix Fetcher to store redirected pages and to store
  protocol-level status. This also should fix NUTCH-273.

* Change default Fetcher behavior not to follow redirects immediately.
  Instead Fetcher will record redirects as new pages to be added to CrawlDb.
  This also partially addresses NUTCH-273.

* Detect and report when Generator creates 0-sized segments.

* Fix Injector to preserve already existing CrawlDatum if the seed list
  being injected also contains such URL.

This development was partially supported by SiteSell Inc.



git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/trunk@490607 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/segment/SegmentReader.java
@@ -95,7 +95,7 @@ public class SegmentReader extends Configured implements Reducer {
       final Path segmentDumpFile = new Path(job.getOutputPath(), name);
 
       // Get the old copy out of the way
-      fs.delete(segmentDumpFile);
+      if (fs.exists(segmentDumpFile)) fs.delete(segmentDumpFile);
 
       final PrintStream printStream = new PrintStream(fs.create(segmentDumpFile));
       return new RecordWriter() {
